requirements.txt
python_vlc==3.0.12118
qt_material==2.8.10
PyQt5==5.15.4
